PIA-689: Manage 1 testapp TCP flow
When a managed application sends traffic on a new socket,
handleNewFlow is called. (just the first time)

A new socket, managed by the network extension process,
is created pointing to the same endpoint and port of the original one.

We start by reading the OUTBOUT traffic of the application
using readTCPFlowData.
We then write the same data to the endpoint via the newly created socket.
(If the endpoint is a netcat on listen, you will see the data)

In readTCPFlowData closure we first write any INBOUND traffic coming
from the new socket back to the application.
We then call readTCPFlowData and block waiting for new OUTBOUND traffic.

Binding and support for multiple connections will be handled on new PRs.