PIA-755: Remove local TCP connection
This connection was a leftover from a previous approach.
Since it is no longer needed, it has been removed.

Also removed other unused settings.

A VPN server remote address is still required during set up.
For now I have hard-coded localhost
(reason for it: our proxy is running locally).
The official docs do not say much, it may also be that the
settings is not used in the context of a NETransparentProxy,
but the API still requires you to have one.