PIA-730: Make TCP flows IO operations non blocking on main thread
This is a first try of how managing multiple flows could
look like.

manageTCPFlow() is detached in a task, to not block the main
thread that needs to be "intercepting" new connections.

The function socket.readData() is also blocking because of recv().
As a workaround, the function and its closure are detached in a
task as well.
When we will refactor the socket class using a library
(probably SwiftNIO), hopefully the functions will support async,
making this workaround useless.