PIA-1042 Ignore LAN and IPv6
Ensure we don't proxy LAN or IPv6. Unfortunately we can't set an IPv6 rule (it seems) so we can only ignore it when we inspect the packets rather than in proxy settings. We can ignore LAN networks in settings, however.

We do not need to explicitly ignore localhost - as localhost seems to be ignored by the proxy by default (in my tests).