PIA-1061: Add multiple UDP flows support using SwiftNIO
We could be reading more than 1 datagram from readDatagrams().
Atm we schedule a new read each time a datagram is successufully written to the flow.
This could result in scheduling a read, when one is already schedule.
This triggers an error.
We can fix it later, but can also keep this behaviour if the error is not closing the flow
or crashing the extension