PIA-1061: Refactor how flow type is handled
Now, the flow type starts as generic and stays so until we actually need
subclasses methods.
There is no difference between TCP or UDP in STProxyProvider, it is all hidden
in the TrafficManager implementation.

Swift does not allow full C++ style run-time polymorphism, that is why
handleFlowIO() has an ugly switch. Not sure we can do better than that :(