PIA-1250 Ensure we mock out ProxySessionFactory in engine tests
Otherwise we end up hitting the real network - not good. Hoewver doing this reduces our test coverage as when we were hitting the real network it was exercising complete code paths, but this isn't good for our tests. I'll add in tests for each component separately to bring the coverage back up later