PIA-1263 Check for a nil NetworkInterface ip
Trace a nil interface ip - we require this ip to exist to bind sockets to it. Before this fix we force unwrapped the ip resulting in a run-time crash, but here we trace it instead and then ignore the flow.

In the future we may block the flow for a vpnOnly flow.

I haven't noticed crashes related to the force unwrapping before, but it's technically possible, so let's guard against it